Fixup paths from app move and brave-extension to brave_extension
