--- components/brave_extension/extension/brave_extension/background.html
@@ -1,6 +1,0 @@
-<!doctype html>
-<html>
-    <head>
-        <script src="/out/brave_extension_background.bundle.js"></script>
-    </head>
-</html>
